make 'got update' bump the base commit ID of unchanged files
This change makes it actually possible to get around commit-time
out-of-dateness by running 'got update'.

The test added with this commit shows that our out-of-dateness
check is currently too simplistic; an update is required between
any two commit operations! It would be better to allow commits to
proceed until a situation arises where file content must be merged.