Print newline to clear the bottom line when tog exits.
When a curses(3)-based application exits, the cursor is positioned
at the start of the bottom line, and the shell will issue the new
prompt there.  However, the bottom line may still contain remnants
of the curses screen.  This is only noticeable on terminals that
do not switch to/from an alternate screen buffer.

ok stsp