fix matching the object ID of the first object in a pack index
An object ID which appears first in a pack index was wrongly ignored by
got_packidx_match_id_str_prefix(). The result was a spurious "object not
found" error when this object ID was passed on the command line.

Reported by jrick on freenode who was lucky enough to this edge case!