fix potential type mismatches between format specifiers and arguments
Cast printf arguments of type time_t and off_t to long long to match
the %lld format specifier on platforms where this might not be the
case.  In parse.y, switch the number variable to long long because
all its interactions are with that type anyway.

ok millert stsp