remove github capabilities workaround from got-send-pack, it is not needed
This workaround was inherited from git9, which claims that Github did
not send capabilities. Protocol traces show that Github's server does
in fact advertise its capabilities. Git protocol capabilities are hidden
behind a NUL byte in the middle of the first refline sent by the server.
This makes them easy to miss when treating a refline as a C string.
I guess this is why Ori concluded that Github didn't send any capabilities.
Or perhaps Github has been fixed since.