fix seek to incorrect offset in the delta base when creating deltas
The stretchblk() function needs to compare data located after the block
which has just been matched. However, upon entry it was resetting the
file pointer of the delta base to the beginning(!) of the block.
The other file is correctly positioned after the block.
In many cases the data won't match and stretchblk() will not stretch
the matched block. But when the data did happen to match this resulted
in a bogus delta, and wrong file contents when the delta was applied.
Fix this by setting the delta base file pointer to end of the block.

Problem reported by naddy after our server refused a pack file which
was sent by 'got send'. I could reproduce the issue by running the
'gotadmin pack' command on a copy of naddy's repository.

ok naddy