regress: fix test failures using date(1)
This patch fixes test failures related the date(1) utility.

In the long term we could add a wrapper function to detect valid options
for date(1) and use the variant which works on the current OS.

For now, this makes some tests pass and shows us where the problems are.

OK thomas.adam