fix histedit_no_op test which was failing randomly
A no-op replayed history ends up having exactly the same commit IDs
if all commits are created at roughly the same moment in time.
There are no content changes involved so if commit timestamps do not
differ then commit hashes will be the same. In which case there is no
fork in history for 'got histedit -l' to display, yet the test was
always expecting a fork in history to be displayed.

Update the test to take this issue into account. The test will now pass
no matter which result is produced by the histedit operation.

Problem found by Lucas who observed that this test was randomly failing.
Patch also provided by Lucas.