stop using the term 'obstructed' ambiguously in the got.1 man page
Specifically, obstructed files are versioned files which should be on disk
but have been replaced on disk by a non-regular file (e.g. by a directory).
The man page used the phrase "deletion was obstructed" which is unrelated
to the above concept and can be expressed as "deletion was prevented" instead.