fix 'got update' of an added + obstructed file
When 'got update' tried to add a new file to the work tree and this
file was obstructed by, say, a directory on disk, the update failed:

  $ got update
  ?  new
  got: new: Is a directory
  $

And the work tree was not updated.

With this commit this situation is properly detected as an obstruction
and the update succeeds:

  $ got update
  ~  new
  Updated to refs/heads/master: c1f85b4938dc4c668a88f13df2b98a520fc077cc
  File paths obstructed by a non-regular file: 1
  $

Extend a corresponding test case to cover this issue.

ok tracey