switch branches during noop rebases
Always update to the specified branch even if the branch is already
rebased (no commits needed to be rebased, and the branch does not need
a fast forward).  With the old behavior of erroring and staying on the
current branch, I sometimes found myself ignoring the error message,
treating it as information and assuming the branch update, and later
on committing on top of origin/main instead of the main branch.

feedback and ok stsp