portable: running tests when shell is dash
Ubuntu's default shell is dash, when using /bin/sh.  The portable nature
of got is such that "$OSTYPE" to determine the host type (linux, bsd,
etc.) is bash-specific and is not part of POSIX.

autotools already provide a mechanism for determining the underlying
platform type, and PLATFORM is already a subst value.

Therefore, let the -portable parts of the regress test-suite use
$PLATFORM to look at the underlying OS type.  The tests themselves
already expect /bin/sh to be POSIX-compliant in all other areas, so
there should be no need to change the #! lines.