portable: support NetBSD
This adds support for NetBSD, using the following external dependencies:

  libuuid
  ncursesw

There is no CI (yet) as enabling that on Cirrus-CI requires some work,
so for now, these builds will be run manually before releases.

The test suite will also need tweaking to be fully supported but the key
tests are supported for now.