portable: improve macos compat with bison/openssl
Some MacOS systems might not have GNU Bison installed, or have openssl
installed.  In such instances the host versions of those can be used,
with a few tweaks around the edges to make them work.

This commit addresses these by:

* Relaxing a HAVE_CRYPTO check for __APPLE__, since the inclusion of the
  machine-specific headers isn't dependent on openssl being installed;
* If the non-GNU version of bison on MacOS is being used, define any
  missing YY* variables, and add a compiler #warning so we log
  something.

This also updates the CI scripts to default to the system versions of
bison and removes openssl as a core dependency -- it's more likely
openssl is going to be installed, so removing it here will mean we can
catch the case for just the MacOS-specific failures, and the inclusion
of openssl on other OSes should cover the rest.

Noticed by Evan Silberman