fix paths stored in pack meta data, improving file deltification
The old code was broken and stored an empty path or filenames, instead
of a repository-relative path. Which means we didn't sort files for
deltification as was intended.
Fixing this provides much better deltas in large pack files written by
gotadmin pack -a. In my test case, pack size changed from 2GB to 1.5GB.

ok op@