store a path hash instead of a verbatim path in pack meta data
This reduces memory use by gotadmin pack. The goal is to sort files
which share a path next to each other for deltification. A hash of
the path is good enough for this purpose and consumes less memory
than a verbatim copy of the path. Git does something similar.

ok op@