fix a bug in findwixt() which caused pack files with missing parent commits
The 'nskip' variable is supposed to reflect commits which are waiting on
the queue and have the 'skip' color. Only increment 'nskip' when adding
such commits to the queue.

Problem observed with got send -T and a tag pointing to a deleted branch.
Test to reproduce the bug written by op@.