fix a segfault in tog diff
The f1 tempfile must always be created. Even if the initial diff has no
objects on the left side, the view can be switched to a different diff
which does have objects on both sides. When that happened, tog crashed.
(To reproduce: open tog in got.git, hit G, hit Enter, hit <)