got patch: use status G for merged files (and document it)
Suggested by stsp@.  This makes possible to tell if 'got patch' used a
merge-base from the repository for the 3-way merge or did a simple
search/replace operation.

While here update the tests to look for the 'G' status and the manpage,
which was also missing the documentation for the 'C' status in got
patch.

ok stsp@