got patch: ignore blobs not found
since diffs are often enclosed in other formats (e.g. emails) we might
parse something and think it's a blob id when it's not.  This should
already happens, but apply_patch is looking for the wrong error due to a
leftover from previous attempts.  Reported by stsp@

while here tweak the test_patch_merge_unknown_blob to also try with a
dummy commit id, as now got-read-patch requires it in order to consider
a blob id.

ok stsp@