portable: regress: set PLATFORM when running direct
PLATFORM is populated when running via `./configure && make` but this
isn't guaranteed if an individual test is run directly, such as
`regress/cmdline/tag.sh`.  In such cases, PLATFORM will be empty, but we
still want to use it.  Since we test for non-linux values, only set
PLATFORM if we're running on Linu so that the correct commands are used.