portable: regress: harden coreutils check
When running tests on non-Linux systems, it was just assumed that GNU
coreutils had been installed.  This meant that the wrapper commands for
date/ln on those systems always had gdate and gln installed.

Rather than assume that, check to see if the GNU coreutils versions as
present, and if not echo a warning.