avoid allocating too many errors in cmd_info
got_error_path uses get_custom_err to get a statically allocated error
allocated on a buffer that's used a ring.  This is nice and simple, but
the drawback is that thees custom errors shouldn't be used "for too
long" because they might get overwritten later.

cmd_info is one offender: change it to store a "simple" error and later
re-use it to construct the per-path error.

OK and tweaks stsp@