tog: fix log 'd' keymap when last commit is displayed
When at the end of the log view such that the currently selected line is
greater than nlines/2, add--don't assign--the page down value to the selected
index so we don't jump up when the user expects tog to scroll down.  While
here, rework log page down logic for a smoother UX in horizontal split mode by
moving the selection cursor down rather than scrolling when 'd' or 'f' is
entered if the last commit is already on screen.

ok stsp@