tog: fix delayed tree index update
If a child blame view has been opened, halfdelay() is reset to a 1 second
refresh rate, so the index in the header lags behind subtree navigation as we
set the index based on the selected tree entry, which isn't updated till that
refresh cycle completes. Don't wait for the selected_entry to update; instead,
compute the index from the first displayed entry and the currently selected
line index.

improved by and ok stsp@