gotwebd: add fcgi_printf
instead of fcgi_gen_response which outputs only a fixed strings provide
a printf-like fcgi_printf: it greatly simplifies the generation of the
HTML pages.

While here also (probably) fix some HTML errors: the output was verified
with the W3C validator and it's correct (in the sense that the tags are
properly closed, there are still some other things the validator
complains about.)

ok/encouragement baseprime@, ok jamsek

Thanks for reading such a boring diff!