gotwebd: fix usage of the commit graph iter
got_get_repo_commits mixes ids allocated on the heap and the one
returned by the iter_next in the same variable, and then even attemp to
free it.

This is both a leak (we loose the pointer to the previously allocated
id) and a possible invalid free since the object id pointer returned by
the graph iterator is not to be passed to free(3).

part of a bigger diff that's ok stsp@