plug leak in the commit graph iterator
We fail to release the memory for the nodes.  To fix it however, we some
consumer of the commit graph iterator need to be corrected: the returned
pointer is safe to be used only up until the next iter_next call;
save a copy it if it's needed afterwards too.

ok stsp@