portable: macos: look for gnu bison harder
When checking for GNU Bison (and presumably other) toolchain commands on
MacOS, ther's two main systems most users use.  Brew, and Ports.

Both of these take different approaches when building software, and
although a Portsfile exists for -portable now, those users who don't use
that directly, or use brew, and try and build -portable from git
directly, will likely be at the mercy of ./configure being able to go
off and find the appropriate tools.

Therefore, ./configure *tries* to be smart about handling this, and will
go and check if brew(1) is installed and try and use that to determine
the likely path for where it puts its software.

Indeed, for MacPorts, the story is very different, and although some
users might therfore use --prefix to ./configure to determine this path,
there is an overloaded assumption between --prefix meaning where the
binaries are to be installed, and --prefix meaning "go and look
elsewhere for other applications".

At some point, a dedicated configure option might be required.

Reported by grey and jamsek.