portable: remove sys/tree.h compat inclusion
libbsd as a library was created some years back to pull together
BSD-specific functionality to be used on non-BSD systems.  When it was
in its infancy, and not being widely adopted across non-BSD systems,
there was often a need to carry around compat versions of
sys/{queue,tree}.h

This is no longer the case -- and the compat code for this has never
worked on systems which lacked a compatible sys/tree.h header anyway, so
remove this, and let the user know that libbsd is needed.

A few people have mentioned this to me.