move functions which open objects into new file object_open_privsep.c
For the future, this will make it possible to provide alternative
implementations of functions now stored in object_open_privsep.c.
This will probably be needed by future gotd(8) which runs inside
a chroot(2) environment and without the "exec" pledge(2) promise,
making it impossible to run libexec helpers on the fly.
Details of this design are not yet settled, but moving functions
into a separate compilation unit won't hurt in any case.