portable: check for sys/tree.h more fully
When checking for sys/tree.h on non-BSD systems, either it will be in
the standard include path (such as on MacOS), or it won't be found,
until libbsd is overlayed.

In cases where we want to find sys/tree.h via libbsd, perform a
AC_CHECK_DECL for a known #define in sys/tree.h -- which will still find
sys/tree.h in either location, or uses compat/tree.h in all other cases.