got patch: handle the removal of binary files
Diffs that remove binary files don't have hunks so got patch would skip
over them, treating that part of the diff as "noise".

Different programs outputs a slightly different diff for this kind of
patches, but the "Binary files ... and /dev/null differ" is usually
shown, so try to match it.

The adedd regress test covers got diff, git diff, and OpenBSD'
/usr/bin/diff.  CVS diffs will currently fail because the guessed file
name will be wrong.  CVS prints the file name in the Index and RCS lines
which got patch currently ignores, and shows an useless path in the
"Binary files /tmp/cvs... and /dev/null differ" line.

Discussed with and ok stsp@