gotwebd: use Content-Disposition for RSS; sets the filename
Since gotwebd serves the RSS from a path without a file name component,
browsers just make up a random string to give the file a name which can
be confusing.  Furthermore, since it's served as application/rss+xml
they try to render the XML (failing.)

Use gotweb_render_content_type_file (which sets the Content-Disposition
HTTP header) to fix both: the feed is now called <repo-name>.rss and set
as attachment (thus saved and not rendered.)  Change the function to
take an optional filename suffix for the occasion.

ok jamsek