fix tempfile handling in gotd's got_object_raw_open()
There was a bug where we reused a file stored in *outfd from a
previous call, resulting in a raw object backed by a file but
with tempfile_idx -1. This then caused bad confusion during
deltification. Fix this by passing tempfd to read_raw() functions
and only setting *outfd in case it is actually required, and
returning tempfd to the repository tempfile pool otherwise.

ok tracey