gotwebd: avoid full history traversal in briefs/commits
This purposefully breaks the 'previous' button in the commits and briefs
page.  It's hard to find the parent of a commit since they can only be
iterated forward.  The way the previous button was generated was to walk
the history from the HEAD down to the specified commit.  This is costly
but more importantly leads to issue when dealing with paths that were
deleted from the repository.

Discussed with stsp and tracey, ok jamsek.