fix missing commits in pack files created with packed object enumeration
got-read-pack forgot to send a tree-enumeration-done message to the
main process if the tree of a given commit had already been traversed.
The main process would then not add the corresponding commit to the
pack file, even though it should be added.

Found while using 'got send' towards gotd in order to populate an
empty repository on the server with non-trivial history, where some
commits always ended up missing due to this bug.

ok tracey