avoid traversing enumerated commits more than once in got-read-pack
Keep track of parent commits that will be processed as part
of looping over the commit queue provided by the main process,
and do not add these commits to the queue again.
Fixes pointless traversal of commits on the queue which will
simply be skipped. The end result is the same either way.

ok tracey