regress for new backout/cherrypick logmsg references
Tests only cover the -lX interface to list and delete logmsg refs created
by previous backout and cherrypick operations. Further tests are needed
for revert and commit coverage to ensure that the correct log messages of
bo/cy commits are loaded into the editor when 'got commit' is run, and that
only the refs used to populate the editor for the new commit are deleted.
Potential printf issues observed by op remain in the diff; noted here as a
reminder to do a sweep at some point in the future as there are other such
instances in cherrypick.sh and possibly elsewhere in regress.

Reviewed as part of previous commit introducing logmsg ref support for
backed-out/cherrypicked commits.

ok stsp@