got: load editor with backout/cherrypick commit log messages
Create new logmsg references used to populate the editor with the log messages
of backed-out and cherrypicked commits when 'got commit' is run.  We use
versioned path state heuristics during revert and commit commands in an effort
to ensure the editor is only populated with the log messages of backed-out
and/or cherrypicked commits involving paths with local changes that are being
committed. This is an implementation of an initial idea and algorithm designed
by stsp.

inputs, improvments and ok stsp@