improve heuristics for prepopulated log messages
Small tweak to cover the case where users specify paths to be committed such
that none of the paths modified by a backout or cherrypick operation are in
the commit. In this case, the logmsg ref will remain in refs/got/worktree to
be used when the user does run a commit involving the affected paths.

ok stsp@