portable: make bison's location configurable
Some systems have bison (yacc) in a non-standard location.  Since
Autotolls is setting the YACC environment variable based on the
AC_PROG_YACC macro, let the user override this if YACC is set in the
environment already.

This issue was found by a user of MacPorts.  I did not get any feedback
from them as to whether this worked, but my own testing indicates this
should work.