gotwebd: fix briefs/tags navigation overlap
Introduces a separate field for the "next" button (now called "More")
and adjusted the CSS/HTML so it's used. Finally, drops the old code used
to handle the pagination for the BRIEFS/COMMITS cases.

Needs a small hack for the time being, setting the action to TAGS.  This
is due how qs->action is changed in gotweb.c while handling the request.

ok jamsek