improve 'got fetch' behaviour when work tree's branch is not on server
Only fetch the work tree's branch if the -b option is not specified.
This keeps -b functional as an override when invoked in a work tree.

Our previous changes did not consider that got.conf is also a source
of lists of branches to fetch, and that -b is supposed to work as an
override of any default behaviour. We were implicitly appending the
work tree's branch as if it was mentioned as an override on the
command line, which was wrong and based on a misunderstanding of
the intended behaviour.

Without -b on the command line we obtain a list of branches to fetch
from got.conf and use this list if it is not empty. The repository's
HEAD will be fetched only if neither the -b option, nor got.conf, nor
a work tree tell us what to fetch.

Make the man page more clear by moving the explanation of the default
behaviour into the main section of 'got fetch', leaving the -a and -b
option descriptions free of such details.

ok jamsek